+++ packages/hono-takibi/src/generators/openapi/components/anyof/generate-anyof-code.ts
@@ -1,6 +1,6 @@
 import type { Schema } from '../../../../types'
 import type { Config } from '../../../../config'
-import { generateZodSchema } from '../../../zod/generate-zod-schema'
+import { generateZodSchema } from '../../../zod/schema/generate-zod-schema'
 import { getRefSchemaName } from '../../../../core/schema/references/get-ref-schema-name'
 import { generateZodUnion } from '../../../zod/generate-zod-union'
