+++ packages/hono-takibi/src/generators/openapi/components/generate-components-code.ts
@@ -1,7 +1,7 @@
 import type { Components } from '../../../types'
 import type { Config } from '../../../config'
-import { generateZodSchemaDefinition } from '../../zod/generate-zod-schema-definition'
-import { generateZodSchema } from '../../../generators/zod/generate-zod-schema'
+import { generateZodSchemaDefinition } from '../../zod/openapi/generate-zod-schema-definition'
+import { generateZodSchema } from '../../zod/schema/generate-zod-schema'
 import { resolveSchemasDependencies } from '../../../core/schema/references/resolve-schemas-dependencies'
 import { generateSchemasExport } from '../paths/generate-schemas-export'
 import { getVariableSchemaNameHelper } from '../../../core/helper/get-variable-schema-name-helper'
