+++ packages/hono-takibi/src/generators/zod/openapi/generate-zod-to-openapi.ts
@@ -1,4 +1,4 @@
-import type { ExampleValue } from '../../types'
+import type { ExampleValue } from '../../../types'
 
 /**
  * Generate OpenAPI example value from Zod example value
